Multiple warning fixes

+++ libovvc/ovutils.c
@@ -39,7 +39,7 @@ ov_log_default(void* ctx, int log_level, const char* log_content, va_list vl)
             type = vvctype;
         }
         fprintf(stderr, "%s", OVLOG_COLORIFY[log_level]);
-        fprintf(stderr, "[%s @ %p] : ", type, (long unsigned)ctx);
+        fprintf(stderr, "[%s @ %p] : ", type, ctx);
         vfprintf(stderr, log_content, vl);
         fprintf(stderr, "%s", RST);
     }
